25.06.02 count2

+++ board_login/250531/pass/back/delete_process.php
@@ -3,18 +3,20 @@
     include ("db_connect_pass.php");
 
     //🔄️フォームから送られてきたデータを変数で受け取る
-    $id = $_GET['id'];
-    $sql = "SELECT * FROM board WHERE id = ?";
+    $id = $_POST['id'];
+    $password = $_POST['password'];
+
+    $sql = "SELECT * FROM board WHERE id = ? AND password = ?";
     $stmt = $conn -> prepare($sql);
-    $stmt -> bind_param("i", $id);    //i: int
+    $stmt -> bind_param("is", $id, $password);    //i: int
     $stmt -> execute();
     $result = $stmt -> get_result();
-    $row = $result -> fetch_assoc();
 
     //❌投稿が存在しない場合は処理を中断
-    if(!$row){
-    die("게시글을 찾을 수 없습니다.");
+    if ($result -> num_rows === 0) {
+        die("비밀번호가 틀렸거나 게시글이 존재하지 않습니다.");
     }
+    $stmt -> close();
 
     //🛍️投稿を削除するSQLを準備
     $sql = "DELETE FROM board WHERE id = ?";
@@ -29,5 +31,6 @@
     }
 
     //🚪データベースから出る
+    $stmt -> close();
     $conn -> close();
 ?>
